refactor: han, cho piece generator 를 인터페이스로 사용하지 않도록 변경
